replace SNAPGraphReader with FastSNAPGraphReader and fix exposition to python
